refactor: move specific validations to better spots
