+++ src/main/java/racingcar/Game.java
@@ -11,12 +11,9 @@ public class Game {
         this.carStorage = carStorage;
     }
 
-    public List<Car> saveCars(List<String> carNames) {
-        List<Car> cars = carNames.stream()
-                .map(Car::new)
-                .toList();
-        updateStorage(cars);
-        return cars;
+    public List<Car> setUp(List<String> cars) {
+        validateCars(cars);
+        return saveCars(cars);
     }
 
     public List<GameProgress> playAll(int trials) {
@@ -31,6 +28,14 @@ public class Game {
         return carStorage.getWinners(leadingCar);
     }
 
+    private List<Car> saveCars(List<String> carNames) {
+        List<Car> cars = carNames.stream()
+                .map(Car::new)
+                .toList();
+        updateStorage(cars);
+        return cars;
+    }
+
     private void updateStorage(List<Car> cars) {
         cars.forEach(carStorage::saveCar);
     }
@@ -42,5 +47,25 @@ public class Game {
                 .toList();
         updateStorage(updatedCars);
         return GameProgress.of(updatedCars);
+    }
+
+    public void validateCars(List<String> cars) {
+        validateListNotEmpty(cars);
+        validateCarDistinct(cars);
+    }
+
+    private static void validateListNotEmpty(List<String> cars) {
+        if (cars.isEmpty()) {
+            throw new IllegalArgumentException("입력된 자동차가 없습니다.");
+        }
+    }
+
+    private static void validateCarDistinct(List<String> cars) {
+        long distinctCars = cars.stream()
+                .distinct()
+                .count();
+        if (cars.size() != distinctCars) {
+            throw new IllegalArgumentException("중복된 이름이 존재합니다.");
+        }
     }
 }
